test(runtime): Consolidate gate/up pipeline matmul_acc test into test_matmul
Move TestPipelineMatmulAccGateUp from the top-level
tests/st/runtime/test_pipeline_matmul_acc.py into the existing matmul topic
file tests/st/runtime/ops/test_matmul.py, alongside the other matmul +
pl.pipeline(stage=2) regression cases. The test method joins the existing
TestMatmulOperations suite. No change to the kernel or expected computation.
